Asset File Loading!!!!!!!!!!!!!!!
here ive added asset file loading, and ive also made linking assets much easier with master builders, you add builders to the master then buildall to build them and link their assets. this is also important for asset file loading,because of delegates! delegates cant be serialized, so i had to create a SerializableAsset class, this stores the variables along with the delegates info, and then once mod assemblies are loaded, the assets are loaded